+++ R/data.R
@@ -1,0 +1,6 @@
+#' List of weighted on/off matrices Z up to p = 4 features
+#'
+#' @format A list of matrices
+"Z_exact"
+
+globalVariables("Z_exact")